+++ scripts/game.py
